Add print option for AccumuloGraphConfiguration JUnit refactoring
